+++ src/index.ts
@@ -1,3 +1,5 @@
 export { Reader as SvfReader } from './svf/reader';
-export { Writer as GltfWriter } from './gltf/writer';
+export { Downloader as SvfDownloader } from './svf/downloader';
 export { Reader as OtgReader } from './otg/reader';
+export { Downloader as OtgDownloader } from './otg/downloader';
+export { Writer as GltfWriter } from './gltf/writer';