+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added
+  - SVF and OTG downloader classes
+
 ## [3.1.2] - 2020-03-12
 
 - Fixed